+++ HISTORY.rst
@@ -16,6 +16,7 @@ Unreleased
 
 * ALCYN-120: Fix missing timestamp kind on export of product when running from cron ESB
 * ALCYN-120: Fix error when locking multiple records on ESB stock export
+* ALCYN-115: Fix saleorder creation when address has ref from partner
 
 **Build**